Added key class and test
